+++ mslicer/src/app/mod.rs
@@ -26,7 +26,7 @@ use crate::{
     },
     windows::{self, Tab},
 };
-use common::{config::SliceConfig, progress::CombinedProgress};
+use common::progress::CombinedProgress;
 use slicer::{format::FormatSliceFile, slicer::Slicer};
 
 pub mod config;
@@ -78,7 +78,7 @@ impl App {
             None => *surface = Tree::new(vec![Tab::Viewport]),
         }
 
-        let slice_config = SliceConfig::default();
+        let slice_config = config.default_slice_config.clone();
         let selected_printer = (config.printers.iter())
             .position(|x| {
                 x.resolution == slice_config.platform_resolution
@@ -263,6 +263,6 @@ fn default_dock_layout(surface: &mut Tree<Tab>) {
 
     let [_old_node, new_node] = surface.split_left(NodeIndex::root(), 0.2, vec![Tab::Models]);
     let [_old_node, new_node] =
-        surface.split_below(new_node, 0.5, vec![Tab::SliceConfig, Tab::Supports]);
-    surface.split_below(new_node, 0.5, vec![Tab::Workspace, Tab::RemotePrint]);
+        surface.split_below(new_node, 0.4, vec![Tab::SliceConfig, Tab::Supports]);
+    surface.split_below(new_node, 0.6, vec![Tab::Workspace, Tab::RemotePrint]);
 }